backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
eventlet==0.35.1
python-engineio==4.9.0
python-socketio==5.11.0
